+++ spec/functional/command/outdated_spec.rb
@@ -1,5 +1,4 @@
 require File.expand_path('../../../spec_helper', __FILE__)
-
 module Pod
   describe Command::Outdated do
     extend SpecHelper::TemporaryRepos
@@ -37,6 +36,65 @@ module Pod
       Command::Outdated.any_instance.stubs(:lockfile).returns(version)
       run_command('outdated', '--no-repo-update')
       UI.output.should.not.include('UIKit')
+    end
+
+    it 'tells the user about outdated pods that can be updated in green' do
+      pod_name = 'BlocksKit'
+
+      current_version_string = mock
+      current_version_string.expects(:green).returns('1.0').once
+      current_version = mock
+      current_version.stubs(:to_s).returns(current_version_string)
+
+      newest_version_string = mock
+      newest_version_string.stubs(:to_s).returns('2.0')
+      newest_version_string.expects(:green).returns('2.0').once
+      newest_version = mock
+      newest_version.stubs(:to_s).returns(newest_version_string)
+      Command::Outdated.any_instance.stubs(:updates).returns([[pod_name, current_version, newest_version, newest_version]])
+      Command::Outdated.any_instance.stubs(:deprecated_pods).returns([])
+
+      run_command('outdated')
+      UI.output.should.include('BlocksKit 1.0 -> 2.0 (latest version 2.0)')
+    end
+
+    it 'tells the user about outdated pods that can not be updated due to version restriction in red' do
+      pod_name = 'BlocksKit'
+
+      version_string = mock
+      version_string.expects(:red).returns('1.0').twice
+      current_version = mock
+      current_version.stubs(:to_s).returns(version_string)
+
+      newest_version = mock
+      newest_version.stubs(:to_s).returns('2.0')
+      Command::Outdated.any_instance.stubs(:updates).returns([[pod_name, current_version, current_version, newest_version]])
+      Command::Outdated.any_instance.stubs(:deprecated_pods).returns([])
+
+      run_command('outdated')
+      UI.output.should.include('BlocksKit 1.0 -> 1.0 (latest version 2.0)')
+    end
+
+    it 'tells the user about outdated pods that can be updated, but not to the latest version in yellow' do
+      pod_name = 'BlocksKit'
+
+      current_version_string = mock
+      current_version_string.expects(:yellow).returns('1.0').once
+      current_version = mock
+      current_version.stubs(:to_s).returns(current_version_string)
+
+      next_version_string = mock
+      next_version_string.expects(:yellow).returns('1.1').once
+      next_version = mock
+      next_version.stubs(:to_s).returns(next_version_string)
+
+      newest_version = mock
+      newest_version.stubs(:to_s).returns('2.0')
+      Command::Outdated.any_instance.stubs(:updates).returns([[pod_name, current_version, next_version, newest_version]])
+      Command::Outdated.any_instance.stubs(:deprecated_pods).returns([])
+
+      run_command('outdated')
+      UI.output.should.include('BlocksKit 1.0 -> 1.1 (latest version 2.0)')
     end
 
     it 'tells the user about deprecated pods' do